pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autodoc"
version = "0.1.0"
description = "AI-powered Python code documentation generator using Gemini and LangChain"
readme = "README.md"
authors = [
    {name = "Abdul Hadi", email = "hadi27767l@gmail.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "jinja2>=3.1.2",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
autodoc = "autodoc.cli:app"

[tool.setuptools]
packages = ["autodoc"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88